your-agent/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
python-telegram-bot>=21.0
chromadb>=0.4.22
pydantic>=2.0.0
